Changed Value from Double to Decimal to match database

+++ branches/RPGMaster/RPGSvc/RPGSvc/Entities/Stat.cs
@@ -11,9 +11,9 @@ namespace RPGSvc.Entities
         public int Id;
         public string Name;
         public string Description;
-        public double Value;
+        public decimal Value;
         
-        public Stat(int id,string name, string description, double value)
+        public Stat(int id,string name, string description, decimal value)
         {
             Id = id;
             Name = name;
